dmaengine: pl330: Set residue in tx_status callback.
Fill txstate.residue with the amount of bytes remaining in the current
transfer if the transfer is not complete.  This will be of particular
use to i2s DMA transfers, providing more accurate hw_ptr values to ASoC.

When setting the residue, account for device to memory transfers.  Do
this by checking both the source and destination addresses against the
active descriptors.  For memory to device the source address will be
incrementing, for device to memory the destination will increment and
for memory to memory both will increment.

Alban Browaeys <prahal@yahoo.com> modified the above:
- to cope with the move of two fields:
the change from pl330_chid to thread in dma_pl330_chan,
and the move of base from pl330_info to  its pl330_dmac container.
- residue meaning from Dylan is the leftover of the current transfer
(that is one period). I make it the leftover  for the whole buffer
as expected by alsa pcm core dmaengine.