HACK: ARM: EXYNOS: Allow for higher CPU frequencies on the ODROID-U2
Also, reenable the L0 frequency level, coincidentally fixing a bug in the
thermal driver.

Those frequencies have been tested on my U2 with the Exynos thermal driver
enabled, no problems were encountered.

NB: the boost frequencies are expected to be tunable via dts in the near
future (if not already). Migrate to dts as soon as possible.