HACK: clk/samsung: revert to no parent for div_mmc_pre4.
Avoid the need to overclock the dw-mshc as a workaround for divider
 recalc ending up calculating a too high divider relation between child
and parent. That is I used 440MHz as when dividing its parent
clocked at 880MHz this made up a recalc divider of 2. With anything lower
the divider was higher and the recalc ended up with too low a clockrate
thus the I/O were failing. With higher a frequency I get a divider too low
thus an even higher overclock.
Without the parent relationship the mshc is clocked at 293.33MHz .