mm/page_alloc: move migratetype recheck logic to __free_one_page()
All the caller of __free_one_page() has same migratetype recheck logic,
so we can move it to __free_one_page(). This reduce line of code and help
future maintenance. This is also preparation step for "mm/page_alloc:
restrict max order of merging on isolated pageblock" which fix the
freepage accouting problem on freepage with more than pageblock order.

Signed-off-by: Joonsoo Kim <iamjoon...@lge.com>