ARM: dts: exynos4412-odroid-common: adjust regulator voltages
Some voltages / voltage ranges differ from the ones that are
used in the Hardkernel vendor kernel. Since the vendor kernel
is heavily tested with respect to system stability, use these
values instead.

TODO: I'm not sure how well these values work for the
ODROID-X, where the SoC is not a Exynos4412 Prime type.