drm/exynos: enable QoS to avoid visual glitches at high resolutions
At the 800MHz MPLL / 400MHz memory bus / 200MHz AXI bus configuration,
the Exynos4412 is unable to reliably scan out a 1080p@60Hz image over HDMI
when the GPU is in use.

Enable some QoS functionality (or rather, simple GPU slowdown) to
avoid this problem. And do the same for VGA for the display modes where
this problem can be seen there.