[HACK] drm/exynos: fix fimc clk disable/unprepare before enable/prepare
pm_rutime_set_active set the state or runtime pm for fimc to active thus
when suspended the fimc_clk_ctrl disable the clocks that were never
activated expicitely.

TODO: check if set_active was correct (ie the hw is already active). If
so find a way to not call the clkd disable/unprepare in firsti runtime
suspend call.